refactor(composeables): fixed the addresses manipulation on the composeables
