+++ packages/composables/src/composables/useUserShipping/index.ts
@@ -4,8 +4,8 @@ import {
   useUserShippingFactory,
   UseUserShippingFactoryParams,
 } from '@vue-storefront/core';
-import { CustomerAddressInput } from '@vue-storefront/magento-api';
 import useUser from '../useUser';
+import { transformUserCreateAddressInput, transformUserUpdateAddressInput } from '../../helpers/userAddressManipulator';
 
 const factoryParams: UseUserShippingFactoryParams<any, any> = {
   provide() {
@@ -16,17 +16,8 @@ const factoryParams: UseUserShippingFactoryParams<any, any> = {
 
   addAddress: async (context: Context, params?) => {
     Logger.debug('[Magento]: addAddress', params.address);
-    const {
-      apartment,
-      ...address
-    } = params.address;
 
-    const createParams: CustomerAddressInput = {
-      ...address,
-      street: [address.street, apartment],
-    };
-
-    const { data } = await context.$magento.api.createCustomerAddress(createParams);
+    const { data } = await context.$magento.api.createCustomerAddress(transformUserCreateAddressInput(params));
 
     return data.createCustomerAddress;
   },
@@ -40,25 +31,8 @@ const factoryParams: UseUserShippingFactoryParams<any, any> = {
 
   updateAddress: async (context: Context, params?) => {
     Logger.debug('[Magento] updateAddress', params);
-    const {
-      apartment,
-      ...address
-    } = params.address;
 
-    const updateParams: CustomerAddressInput = {
-      ...address,
-      street: [address.street, apartment],
-    };
-
-    const updateAddressParams: {
-      addressId: number;
-      input: CustomerAddressInput;
-    } = {
-      addressId: params.address.id,
-      input: updateParams,
-    };
-
-    const { data } = await context.$magento.api.updateCustomerAddress(updateAddressParams);
+    const { data } = await context.$magento.api.updateCustomerAddress(transformUserUpdateAddressInput(params));
 
     return data.updateCustomerAddress;
   },
@@ -76,29 +50,8 @@ const factoryParams: UseUserShippingFactoryParams<any, any> = {
 
   setDefaultAddress: async (context: Context, params) => {
     Logger.debug('[Magento] setDefaultAddress');
-    const {
-      apartment,
-      ...address
-    } = params.address;
 
-    const addressParams: CustomerAddressInput = {
-      ...address,
-      street: [address.street, apartment],
-    };
-
-    const customerAddressParams: {
-      addressId: number;
-      input: CustomerAddressInput;
-    } = {
-      // @ts-ignore
-      addressId: params.address.id,
-      input: {
-        ...addressParams,
-        default_shipping: true,
-      },
-    };
-
-    const { data } = await context.$magento.api.updateCustomerAddress(customerAddressParams);
+    const { data } = await context.$magento.api.updateCustomerAddress(transformUserUpdateAddressInput(params));
 
     return data.updateCustomerAddress;
   },
